update install script

+++ scripts/install.sh
@@ -22,11 +22,12 @@ install_third_party() {
   cmake .. -DCMAKE_INSTALL_PREFIX=$DEP_BUILD_DIR -DBUILD_SHARED_LIBS=ON
   make -j2
   make install
+  ldconfig
   cd $THIRD_PARTY_DIR
 
   # 2. glog:
   cd glog
-  mkdir build; cd build
+  mkdir -p build && cd build
   cmake .. -DCMAKE_INSTALL_PREFIX=$DEP_BUILD_DIR -DBUILD_SHARED_LIBS=ON -Dgflags_DIR=$DEP_BUILD_DIR
   make -j2
   make install
@@ -34,38 +35,48 @@ install_third_party() {
 
   # 3. snappy:
   cd snappy
-  mkdir build; cd build
+  mkdir -p build && cd build
   cmake ..
+  cd $THIRD_PARTY_DIR
+
+  # 4. googletest:
+  cd googletest
+  mkdir -p build && cd build
+  cmake .. -DCMAKE_INSTALL_PREFIX=$DEP_BUILD_DIR -DBUILD_SHARED_LIBS=ON
+  make -j
+  make install
+  cd $THIRD_PARTY_DIR
 }
 
 # Download Pin - A Binary Instrumentation Tool
-download_intel_pin()
-{
+download_intel_pin() {
   cd "$THIRD_PARTY_DIR"
 
-  wget http://software.intel.com/sites/landingpage/pintool/downloads/pin-3.16-98275-ge0db48c31-gcc-linux.tar.gz
-  tar -xvf pin-3.16-98275-ge0db48c31-gcc-linux.tar.gz
-  ln -s "$THIRD_PARTY_DIR"/pin-3.16-98275-ge0db48c31-gcc-linux "$ROOT"/pin 
+  if [ ! -e "pin-3.16-98275-ge0db48c31-gcc-linux.tar.gz" ]
+  then
+    wget http://software.intel.com/sites/landingpage/pintool/downloads/pin-3.16-98275-ge0db48c31-gcc-linux.tar.gz
+    tar -xvf pin-3.16-98275-ge0db48c31-gcc-linux.tar.gz
+    ln -s "$THIRD_PARTY_DIR"/pin-3.16-98275-ge0db48c31-gcc-linux "$ROOT"/pin
+  fi
 }
 
-install_mcsim_frontend()
-{
+install_mcsim_frontend() {
   cd "$ROOT"/Pthread
 
   make clean PIN_ROOT="$ROOT"/pin
-  make PIN_ROOT="$(pwd)"/../pin -j
+  make PIN_ROOT="$ROOT"/pin obj-intel64/mypthreadtool.so -j4
+  make PIN_ROOT="$ROOT"/pin obj-intel64/libmypthread.a
 }
 
-install_mcsim_frontend_debug()
-{
+install_mcsim_frontend_debug() {
   cd "$ROOT"/Pthread
 
   make clean PIN_ROOT="$ROOT"/pin
-  make DEBUG=1 PIN_ROOT="$(pwd)"/../pin -j
+  make DEBUG=1 PIN_ROOT="$ROOT"/pin obj-intel64/mypthreadtool.so -j4
+  make DEBUG=1 PIN_ROOT="$ROOT"/pin obj-intel64/libmypthread.a
 }
 
-install_mcsim_backend()
-{
+install_mcsim_backend() {
   # McSim Back-end Build
   cd "$ROOT"/McSim
   rm -rf build
@@ -74,24 +85,21 @@ install_mcsim_backend()
   cmake .. && make -j
 }
 
-install_mcsim_backend_debug()
-{
+install_mcsim_backend_debug() {
   # McSim Back-end Build
   cd "$ROOT"/McSim
   rm -rf debug
-   
+
   mkdir debug && cd debug
   cmake .. -DCMAKE_BUILD_TYPE=Debug && make -j
 }
 
-install_tracegen()
-{
+install_tracegen() {
   cd "$ROOT"/TraceGen
   make PIN_ROOT="$ROOT"/pin obj-intel64/tracegen.so -j
 }
 
-build_stream_example()
-{
+build_stream_example() {
   cd "$ROOT"/McSim/stream
   make clean && make -j
 }
@@ -111,13 +119,11 @@ unset C_INCLUDE_PATH
 unset PKG_CONFIG_PATH
 
 install_third_party
-
 download_intel_pin
-
 install_mcsim_frontend
-
 install_mcsim_backend
-
 install_tracegen
 
-build_stream_example
+#install_mcsim_frontend_debug
+#install_mcsim_backend_debug
+#build_stream_example
